requirements.txt
Flask
finch-api
python-dotenv